fix: fsd, 데이터 무존재시 ui
