+++ src/entities/attendance/model/index.ts
@@ -1,3 +1,4 @@
 export * from "./mappers";
-
 export * from "./constants";
+export * from "./types";
+export * from "./transformers";
