+++ src/page-layer/attendance/Attendance.module.scss
@@ -37,28 +37,19 @@
   }
 }
 
-.errorContainer {
+.emptyState {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   gap: var(--spacing-16);
-  padding: var(--spacing-32);
-  min-height: 200px;
+  padding: var(--spacing-64) var(--spacing-32);
+  min-height: 300px;
 }
 
-.errorIcon {
-  width: 24px;
-  height: 24px;
-  color: var(--color-system-error);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.errorMessage {
+.emptyMessage {
   font-size: var(--font-size-base);
-  color: var(--color-system-error);
+  color: var(--foreground-secondary);
   text-align: center;
-  font-weight: var(--font-weight-semibold);
+  font-weight: var(--font-weight-medium);
 }
